Extract proposal memo service operations
