+++ src/app/services/proposal_service.py
@@ -14,13 +14,6 @@ from app.contracts.proposals import (
     ProposalLineageEnvelopeResponse,
     ProposalListData,
     ProposalListEnvelopeResponse,
-    ProposalMemoAiCommentaryEnvelopeResponse,
-    ProposalMemoEnvelopeResponse,
-    ProposalMemoLineageEnvelopeResponse,
-    ProposalMemoProjectionEnvelopeResponse,
-    ProposalMemoReplayEvidenceEnvelopeResponse,
-    ProposalMemoReportPackageEnvelopeResponse,
-    ProposalMemoReviewEnvelopeResponse,
     ProposalNarrativeReviewEnvelopeResponse,
     ProposalReportRequestEnvelopeResponse,
     ProposalSimulateResponse,
@@ -33,6 +26,7 @@ from app.contracts.proposals import (
     ProposalWorkflowEventsEnvelopeResponse,
 )
 from app.services.advisory_client_protocols import ProposalClient
+from app.services.proposal_memo_service import ProposalMemoServiceMixin
 from app.services.upstream_envelope import (
     build_gateway_envelope,
     build_typed_gateway_envelope,
@@ -60,7 +54,7 @@ def _normalize_proposal_context_payload(
     return normalized
 
 
-class ProposalService:
+class ProposalService(ProposalMemoServiceMixin):
     def __init__(self, advise_client: ProposalClient):
         self._advise_client = advise_client
 
@@ -603,196 +597,6 @@ class ProposalService:
         )
         self._raise_for_upstream_error(upstream_status, upstream_payload)
         return self._opaque_envelope(correlation_id, upstream_payload)
-
-    async def create_proposal_memo(
-        self,
-        proposal_id: str,
-        version_no: int,
-        body: dict[str, Any],
-        idempotency_key: str,
-        correlation_id: str,
-    ) -> ProposalMemoEnvelopeResponse:
-        upstream_status, upstream_payload = await self._advise_client.create_proposal_memo(
-            proposal_id=proposal_id,
-            version_no=version_no,
-            body=body,
-            idempotency_key=idempotency_key,
-            correlation_id=correlation_id,
-        )
-        self._raise_for_upstream_error(upstream_status, upstream_payload)
-        return build_gateway_envelope(
-            ProposalMemoEnvelopeResponse,
-            correlation_id=correlation_id,
-            upstream_payload=upstream_payload,
-        )
-
-    async def get_proposal_memo(
-        self,
-        proposal_id: str,
-        version_no: int,
-        correlation_id: str,
-    ) -> ProposalMemoEnvelopeResponse:
-        upstream_status, upstream_payload = await self._advise_client.get_proposal_memo(
-            proposal_id=proposal_id,
-            version_no=version_no,
-            correlation_id=correlation_id,
-        )
-        self._raise_for_upstream_error(upstream_status, upstream_payload)
-        return build_gateway_envelope(
-            ProposalMemoEnvelopeResponse,
-            correlation_id=correlation_id,
-            upstream_payload=upstream_payload,
-        )
-
-    async def get_proposal_memo_projection(
-        self,
-        proposal_id: str,
-        version_no: int,
-        audience: str | None,
-        correlation_id: str,
-    ) -> ProposalMemoProjectionEnvelopeResponse:
-        upstream_status, upstream_payload = await self._advise_client.get_proposal_memo_projection(
-            proposal_id=proposal_id,
-            version_no=version_no,
-            audience=audience,
-            correlation_id=correlation_id,
-        )
-        self._raise_for_upstream_error(upstream_status, upstream_payload)
-        return build_gateway_envelope(
-            ProposalMemoProjectionEnvelopeResponse,
-            correlation_id=correlation_id,
-            upstream_payload=upstream_payload,
-        )
-
-    async def review_proposal_memo(
-        self,
-        proposal_id: str,
-        version_no: int,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> ProposalMemoReviewEnvelopeResponse:
-        upstream_status, upstream_payload = await self._advise_client.review_proposal_memo(
-            proposal_id=proposal_id,
-            version_no=version_no,
-            body=body,
-            idempotency_key=idempotency_key,
-            correlation_id=correlation_id,
-        )
-        self._raise_for_upstream_error(upstream_status, upstream_payload)
-        return build_gateway_envelope(
-            ProposalMemoReviewEnvelopeResponse,
-            correlation_id=correlation_id,
-            upstream_payload=upstream_payload,
-        )
-
-    async def record_proposal_memo_report_package_event(
-        self,
-        proposal_id: str,
-        version_no: int,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> ProposalEnvelopeResponse:
-        (
-            upstream_status,
-            upstream_payload,
-        ) = await self._advise_client.record_proposal_memo_report_package_event(
-            proposal_id=proposal_id,
-            version_no=version_no,
-            body=body,
-            idempotency_key=idempotency_key,
-            correlation_id=correlation_id,
-        )
-        self._raise_for_upstream_error(upstream_status, upstream_payload)
-        return self._opaque_envelope(correlation_id, upstream_payload)
-
-    async def request_proposal_memo_report_package(
-        self,
-        proposal_id: str,
-        version_no: int,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> ProposalMemoReportPackageEnvelopeResponse:
-        (
-            upstream_status,
-            upstream_payload,
-        ) = await self._advise_client.request_proposal_memo_report_package(
-            proposal_id=proposal_id,
-            version_no=version_no,
-            body=body,
-            idempotency_key=idempotency_key,
-            correlation_id=correlation_id,
-        )
-        self._raise_for_upstream_error(upstream_status, upstream_payload)
-        return build_gateway_envelope(
-            ProposalMemoReportPackageEnvelopeResponse,
-            correlation_id=correlation_id,
-            upstream_payload=upstream_payload,
-        )
-
-    async def request_proposal_memo_ai_commentary(
-        self,
-        proposal_id: str,
-        version_no: int,
-        body: dict[str, Any],
-        idempotency_key: str | None,
-        correlation_id: str,
-    ) -> ProposalMemoAiCommentaryEnvelopeResponse:
-        (
-            upstream_status,
-            upstream_payload,
-        ) = await self._advise_client.request_proposal_memo_ai_commentary(
-            proposal_id=proposal_id,
-            version_no=version_no,
-            body=body,
-            idempotency_key=idempotency_key,
-            correlation_id=correlation_id,
-        )
-        self._raise_for_upstream_error(upstream_status, upstream_payload)
-        return build_gateway_envelope(
-            ProposalMemoAiCommentaryEnvelopeResponse,
-            correlation_id=correlation_id,
-            upstream_payload=upstream_payload,
-        )
-
-    async def get_proposal_memo_lineage(
-        self,
-        proposal_id: str,
-        correlation_id: str,
-    ) -> ProposalMemoLineageEnvelopeResponse:
-        upstream_status, upstream_payload = await self._advise_client.get_proposal_memo_lineage(
-            proposal_id=proposal_id,
-            correlation_id=correlation_id,
-        )
-        self._raise_for_upstream_error(upstream_status, upstream_payload)
-        return build_gateway_envelope(
-            ProposalMemoLineageEnvelopeResponse,
-            correlation_id=correlation_id,
-            upstream_payload=upstream_payload,
-        )
-
-    async def get_proposal_memo_replay_evidence(
-        self,
-        proposal_id: str,
-        version_no: int,
-        correlation_id: str,
-    ) -> ProposalMemoReplayEvidenceEnvelopeResponse:
-        (
-            upstream_status,
-            upstream_payload,
-        ) = await self._advise_client.get_proposal_memo_replay_evidence(
-            proposal_id=proposal_id,
-            version_no=version_no,
-            correlation_id=correlation_id,
-        )
-        self._raise_for_upstream_error(upstream_status, upstream_payload)
-        return build_gateway_envelope(
-            ProposalMemoReplayEvidenceEnvelopeResponse,
-            correlation_id=correlation_id,
-            upstream_payload=upstream_payload,
-        )
 
     async def _record_approval(
         self,
